+++ tests/Feature/AccountTest.php
@@ -81,63 +81,4 @@ class AccountTest extends AbstractTestCase
             'lock_version' => 0,
         ]);
     }
-
-    /**
-     * 正常系のテスト
-     * ログインできること
-     */
-    public function testSuccessLogin()
-    {
-        $accounts = factory(Account::class)->create();
-        $accounts->each(function ($account) {
-            factory(QiitaAccount::class)->create(['account_id' => $account->id]);
-            factory(AccessToken::class)->create(['account_id' => $account->id]);
-            factory(LoginSession::class)->create(['account_id' => $account->id]);
-        });
-
-        $permanentId = '1'; // factoryにて登録済みのQiitaアカウントのID
-        $accessToken = 'login0593b2655e9568f144fb1826342292f5c6b7d406fda00577b8d1530d8a5';
-
-        $jsonResponse = $this->postJson(
-            '/api/accounts',
-            [
-                'permanentId' => $permanentId,
-                'accessToken' => $accessToken
-            ]
-        );
-
-        $responseObject = json_decode($jsonResponse->content());
-
-        // 実際にJSONResponseに期待したデータが含まれているか確認する
-        $accountId = 3; // RefreshDatabaseが実行されてautoincrementの値はリセットされない
-        $expectedEmbedded = [
-            'sessionId' => $responseObject->_embedded->sessionId
-        ];
-        $jsonResponse->assertJson(['accountId' => $accountId]);
-        $jsonResponse->assertJson(['_embedded' => $expectedEmbedded]);
-        $jsonResponse->assertStatus(201);
-
-        // DBのテーブルに期待した形でデータが入っているか確認する
-        $idSequence = 3; // RefreshDatabaseが実行されてautoincrementの値はリセットされない
-
-        $this->assertDatabaseHas('accounts_qiita_accounts', [
-            'id'               => $idSequence,
-            'account_id'       => $accountId,
-            'qiita_account_id' => $permanentId,
-            'lock_version'     => 0,
-        ]);
-
-        $this->assertDatabaseHas('accounts_access_tokens', [
-            'id'           => $idSequence,
-            'account_id'   => $accountId,
-            'access_token' => $accessToken,
-            'lock_version' => 0,
-        ]);
-
-        $this->assertDatabaseHas('login_sessions', [
-            'id'           => $responseObject->_embedded->sessionId,
-            'account_id'   => $accountId,
-            'lock_version' => 0,
-        ]);
-    }
 }
